% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textlinedetector.R
\name{lines.textlines}
\alias{lines.textlines}
\title{Extract the polygons of the textlines}
\usage{
\method{lines}{textlines}(x, image, crop = TRUE, channels = c("bgr", "gray"), ...)
}
\arguments{
\item{x}{an object of class \code{textlines} as returned by \code{\link{image_textlines_astar}} or \code{\link{image_textlines_flor}}}

\item{image}{an object of class magick-image}

\item{crop}{extract only the bounding box of the polygon of the text lines}

\item{channels}{either 'bgr' or 'gray' to work on the colored data or on binary greyscale data}

\item{...}{further arguments passed on}
}
\value{
the object \code{x} where element \code{textlines} is replaced with the extracted polygons of text lines
}
\description{
Extract the polygons of the textlines as a cropped rectangular image containing the image content of the line segmented polygon
}
\examples{
## See the examples in ?image_textlines_astar or ?image_textlines_flor
}
