% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deslant.R
\name{ocv_deslant}
\alias{ocv_deslant}
\title{Deslant images by putting cursive text upright}
\usage{
ocv_deslant(x, bgcolor = 255, lower_bound = -1, upper_bound = 1)
}
\arguments{
\item{x}{an object of class opencv-image with pixel values between 0 and 255}

\item{bgcolor}{integer value with the background color to use to fill the gaps of the sheared image that is returned. Defaults to white: 255}

\item{lower_bound}{lower bound of shear values. Defaults to -1}

\item{upper_bound}{upper bound of shear values. Defaults to 1}
}
\value{
an object of class opencv-image with the deslanted image
}
\description{
This algorithm sets handwritten text in images upright by removing cursive writing style. 
One can use it as a preprocessing step for handwritten text recognition.
}
\examples{
\dontshow{
if(require(opencv)) \{
}
library(opencv)
library(image.textlinedetector)
path <- system.file(package = "image.textlinedetector", "extdata", "cursive.png")
img  <- ocv_read(path)
img  <- ocv_grayscale(img)
img
up   <- ocv_deslant(img)
up
\dontshow{
\}
# End of main if statement running only if the required packages are installed
}
}
